refactor(search applications): rename engines component folder
