+++ x-pack/plugins/enterprise_search/public/applications/applications/components/engine/add_indices_flyout.tsx
@@ -33,7 +33,7 @@ import {
   IndicesSelectComboBox,
   IndicesSelectComboBoxOption,
   indexToOption,
-} from '../engines/components/indices_select_combobox';
+} from '../search_applications/components/indices_select_combobox';
 
 import { AddIndicesLogic } from './add_indices_logic';
 import { EngineViewLogic } from './engine_view_logic';
