+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/components/empty_search_applications_prompt.tsx
