+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/components/indices_select_combobox.tsx
