+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/components/tables/search_applications_table.tsx
