+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/create_search_application_logic.test.ts
