+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/delete_search_application_modal.tsx
