+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/search_applications_router.tsx
