+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/types.ts
